--- src/event/events.controller.spec.ts
@@ -1,19 +1,0 @@
-import type { TestingModule } from '@nestjs/testing';
-import { Test } from '@nestjs/testing';
-import { EventController } from './events.controller';
-
-describe('EventController', () => {
-  let controller: EventController;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      controllers: [EventController],
-    }).compile();
-
-    controller = module.get<EventController>(EventController);
-  });
-
-  it('should be defined', () => {
-    expect(controller).toBeDefined();
-  });
-});
