--- src/event/events.service.spec.ts
@@ -1,18 +1,0 @@
-import { Test, TestingModule } from '@nestjs/testing';
-import { EventService } from './events.service';
-
-describe('EventService', () => {
-  let service: EventService;
-
-  beforeEach(async () => {
-    const module: TestingModule = await Test.createTestingModule({
-      providers: [EventService],
-    }).compile();
-
-    service = module.get<EventService>(EventService);
-  });
-
-  it('should be defined', () => {
-    expect(service).toBeDefined();
-  });
-});
