+++ src/events/entities/event.entity.ts
@@ -1,0 +1,41 @@
+import { User } from '@src/users/entities';
+import {
+  Column,
+  CreateDateColumn,
+  Entity,
+  JoinColumn,
+  ManyToOne,
+  PrimaryGeneratedColumn,
+  UpdateDateColumn,
+} from 'typeorm';
+
+@Entity()
+export class Event {
+  @PrimaryGeneratedColumn('uuid')
+  id: string;
+
+  @ManyToOne(() => User, (user) => user.id)
+  @JoinColumn({ name: 'user_id' })
+  user_id: User;
+
+  @Column()
+  title: string;
+
+  @Column({ nullable: true })
+  place: string;
+
+  @Column({ nullable: true })
+  description: string;
+
+  @Column()
+  started_at: Date;
+
+  @Column()
+  ended_at: Date;
+
+  @CreateDateColumn()
+  created_at: Date;
+
+  @UpdateDateColumn()
+  update_at: Date;
+}
