+++ src/events/index.ts
@@ -1,5 +1,4 @@
 export * from './dto';
-export * from './entities';
 export * from './events.controller';
 export * from './events.module';
 export * from './events.service';
