+++ src/users/entities/index.ts
@@ -1,2 +1,1 @@
-// eslint-disable-next-line import/no-cycle
 export * from './user.entity';
