+++ src/users/entities/user.entity.ts
@@ -1,14 +1,14 @@
+import type { Event } from '@src/events/entities';
 import {
   Column,
   CreateDateColumn,
   Entity,
-  ManyToOne,
   OneToMany,
   PrimaryGeneratedColumn,
   UpdateDateColumn,
 } from 'typeorm';
-// eslint-disable-next-line import/no-cycle
-import { Event } from '@src/event/entities';
+
+type EventTypes = Event;
 
 @Entity()
 export class User {
@@ -33,6 +33,6 @@ export class User {
   @Column({ select: false, nullable: true })
   session: string;
 
-  @OneToMany((type) => Event, (event) => event.user)
-  events: Event[];
+  @OneToMany('Event', 'user_id')
+  events: EventTypes;
 }
