+++ src/users/index.ts
@@ -1,6 +1,5 @@
 /* eslint-disable import/no-cycle */
 export * from './dto';
-export * from './entities';
 export * from './users.controller';
 export * from './users.module';
 export * from './users.service';
