+++ src/users/users.module.ts
@@ -1,8 +1,9 @@
 /* eslint-disable import/no-cycle */
 import { Module, forwardRef } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { AuthConfigModule } from '@src/config';
 import { AuthModule } from '@src/auth';
+import { AuthConfigModule } from '@src/config';
+import { Event } from '@src/events/entities';
 import { User } from './entities';
 import { UsersController } from './users.controller';
 import { UsersService } from './users.service';
@@ -10,7 +11,7 @@ import { UsersService } from './users.service';
 @Module({
   imports: [
     forwardRef(() => AuthModule),
-    TypeOrmModule.forFeature([User]),
+    TypeOrmModule.forFeature([User, Event]),
     AuthConfigModule,
   ],
   providers: [UsersService],
