+++ src/users/users.service.ts
@@ -4,7 +4,6 @@ import { AuthConfigService } from '@src/config';
 import * as bcrypt from 'bcrypt';
 import { Repository } from 'typeorm';
 import type { CreateUserDto, UpdateUserDto } from './dto';
-// eslint-disable-next-line import/no-cycle
 import { User } from './entities';
 
 @Injectable()
